Rename class files to meet convention of lowercase names
